server/requirements.txt
Pillow==9.3.0
bcrypt==3.1.2
python_dateutil==2.8.2
SQLAlchemy==1.4.25
